refactoring

+++ eventscan/src/main/java/io/lastwill/eventscan/services/handlers/ContractEventHandler.java
@@ -5,7 +5,6 @@ import io.lastwill.eventscan.events.contract.*;
 import io.lastwill.eventscan.messages.*;
 import io.lastwill.eventscan.repositories.ProductRepository;
 import io.lastwill.eventscan.services.BalanceProvider;
-import io.lastwill.eventscan.services.EventParser;
 import io.lastwill.eventscan.services.ExternalNotifier;
 import io.lastwill.eventscan.services.handlers.events.TransferOwnershipHandler;
 import lombok.extern.slf4j.Slf4j;
@@ -55,7 +54,7 @@ public class ContractEventHandler {
                 balanceProvider.getBalanceAsync(
                         event.getNetworkType(),
                         event.getContract().getAddress(),
-                        event.getBlock().getNumber().longValue()
+                        event.getBlock().getNumber()
                 )
                         .thenAccept(balance -> {
                             try {
